android-10.0.0_r2

+++ external/cn-cbor/src/cn-cbor.c
@@ -52,7 +52,7 @@ static double decode_half(int half) {
 /* Fix these if you can't do non-aligned reads */
 #define ntoh8p(p) (*(unsigned char*)(p))
 #define ntoh16p(p) (ntohs(*(unsigned short*)(p)))
-#define ntoh32p(p) (ntohl(*(unsigned long*)(p)))
+#define ntoh32p(p) (ntohl(*(uint32_t*)(p)))
 static uint64_t ntoh64p(unsigned char *p) {
   uint64_t ret = ntoh32p(p);
   ret <<= 32;
